# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthProvider": {
        "enum": [
          "google",
          "apple",
          "email"
        ],
        "title": "AuthProvider",
        "type": "string"
      },
...
      "DemoLoginRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "DemoLoginRequest",
        "type": "object"
      },
      "EmailCodeRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "EmailCodeRequest",
        "type": "object"
      },
      "EmailCodeVerify": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email",
          "code"
        ],
        "title": "EmailCodeVerify",
        "type": "object"
      },
...
      "ProblemOut": {
        "description": "The error body, declared so it reaches the generated contract.\n\nRendered by an exception handler rather than returned from a route, so\nFastAPI cannot infer it; app/api/router.py attaches it to every operation\nas the default response. problem_response() below builds from this model,\nwhich is what keeps the declaration and the wire body the same shape.",
        "properties": {
          "code": {
            "description": "Stable machine code the client maps to a failure state.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code",
          "type"
        ],
        "title": "ProblemOut",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "auth_provider": {
            "$ref": "#/components/schemas/AuthProvider"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "home_timezone": {
            "title": "Home Timezone",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "auth_provider",
          "created_at",
          "home_timezone"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "UserUpdateIn": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "home_timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Timezone"
          }
        },
        "title": "UserUpdateIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Generated from the FastAPI routers by backend/scripts/dump_openapi.py. Do not edit: run the script. This describes what the server actually serves and is what the frontend's types are generated from. The hand-written docs/openapi.yaml is the design artifact and may declare endpoints ahead of the code; backend/scripts/check_openapi_drift.py holds the two together. See docs/adr/004.",
    "title": "TravelWell API (served)",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/demo": {
      "post": {
        "description": "Mint a fresh demo account pre-populated with the demo trips.\n\nPer-tap accounts keep testers isolated (mutations are real and would\ncollide on a shared account) and build the dates fresh, so the demo never\ngoes stale. The optional name is just a label, not a credential: the\naccount lives only in this browser's session cookie.\n\nThe email is the one thing this route owns; everything else about the\naccount is content, and lives in app/services/demo_user/data.py.",
        "operationId": "demo_login_api_v1_auth_demo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DemoLoginRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Demo Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/email-code": {
      "post": {
        "operationId": "request_email_code_api_v1_auth_email_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Request Email Code",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/email-code/verify": {
      "post": {
        "operationId": "verify_email_code_api_v1_auth_email_code_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailCodeVerify"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Verify Email Code",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/oauth/{provider}/callback": {
      "get": {
        "operationId": "oauth_callback_api_v1_auth_oauth__provider__callback_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Oauth Callback",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/oauth/{provider}/start": {
      "get": {
        "operationId": "start_oauth_api_v1_auth_oauth__provider__start_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Start Oauth",
        "tags": [
          "auth"
        ]
      }
    },
    "/me": {
      "get": {
        "operationId": "get_me_api_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Get Me",
        "tags": [
          "auth"
        ]
      },
      "patch": {
        "description": "Fields on users itself; preferences live on PATCH /me/preferences.",
        "operationId": "update_me_api_v1_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Update Me",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}